updates and add test and clean
