+++ examples/test-site/src/data/pages/filter-item/tags.json
@@ -1,8 +1,8 @@
 {
   "tags": [
     {
-      "id": "4db4e4f8-4828-494f-9d74-f20832b554e6",
-      "name": "Bananas"
+      "id": "6eb99247-beb7-4750-8b1d-918e8fbf2a9b",
+      "name": "bbbb"
     }
   ]
 }
